rename PPW migration files
